ready to move on

+++ src/ddl/ddl_small_bank.sql
@@ -7,14 +7,14 @@ CREATE INDEX IDX_ACCOUNTS_NAME ON ACCOUNTS (name);
 
 CREATE TABLE SAVINGS (
     custid      BIGINT      NOT NULL,
-    bal         DOUBLE       NOT NULL,
+    bal         BIGINT       NOT NULL,
     PRIMARY KEY (custid),
     FOREIGN KEY (custid) REFERENCES ACCOUNTS (custid)
 );
 
 CREATE TABLE CHECKING (
     custid      BIGINT      NOT NULL,
-    bal         DOUBLE       NOT NULL,
+    bal         BIGINT       NOT NULL,
     CONSTRAINT pk_checking PRIMARY KEY (custid),
     FOREIGN KEY (custid) REFERENCES ACCOUNTS (custid)
 );
